mave code to gammapy.datasets.actors
