+++ gammapy/datasets/__init__.py
@@ -1,9 +1,9 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from gammapy.utils.registry import Registry
-from .core import Dataset, Datasets, DatasetsActor
+from .core import Dataset, Datasets
 from .flux_points import FluxPointsDataset
 from .io import OGIPDatasetReader, OGIPDatasetWriter
-from .map import MapDataset, MapDatasetActor, MapDatasetOnOff, create_map_dataset_geoms
+from .map import MapDataset, MapDatasetOnOff, create_map_dataset_geoms
 from .simulate import MapDatasetEventSampler
 from .spectrum import SpectrumDataset, SpectrumDatasetOnOff
 
@@ -24,10 +24,8 @@ __all__ = [
     "Dataset",
     "DATASET_REGISTRY",
     "Datasets",
-    "DatasetsActor",
     "FluxPointsDataset",
     "MapDataset",
-    "MapDatasetActor",
     "MapDatasetEventSampler",
     "MapDatasetOnOff",
     "OGIPDatasetWriter",
@@ -35,3 +33,12 @@ __all__ = [
     "SpectrumDataset",
     "SpectrumDatasetOnOff",
 ]
+
+
+try:
+    from .actors import DatasetsActor, MapDatasetActor
+
+    __all__.extend(["MapDatasetActor", "DatasetsActor"])
+
+except ModuleNotFoundError:
+    pass
