+++ gammapy/datasets/core.py
@@ -8,7 +8,6 @@ import logging
 import numpy as np
 from astropy import units as u
 from astropy.table import Table, vstack
-import ray
 from gammapy.data import GTI
 from gammapy.modeling.models import DatasetModels, Models
 from gammapy.utils.scripts import make_name, make_path, read_yaml, write_yaml
@@ -16,7 +15,7 @@ from gammapy.utils.scripts import make_name, make_path, read_yaml, write_yaml
 log = logging.getLogger(__name__)
 
 
-__all__ = ["Dataset", "Datasets", "DatasetsActor"]
+__all__ = ["Dataset", "Datasets"]
 
 
 class Dataset(abc.ABC):
@@ -597,67 +596,3 @@ class Datasets(collections.abc.MutableSequence):
 
     def __len__(self):
         return len(self._datasets)
-
-
-class DatasetsActor(Datasets):
-    """A modified Dataset collection for parallel evaluation using ray actors.
-    Fore now only available if composed only of MapDataset.
-
-    Parameters
-    ----------
-    datasets : `Datasets`
-        Datasets
-    """
-
-    def __init__(self, datasets=None):
-        from .map import MapDataset, MapDatasetActor
-
-        if datasets is not None:
-            actors = []
-            datasets_list = []
-            while datasets:
-                d0 = datasets[0]
-                actors.append(MapDatasetActor.remote(d0))
-                datasets_list.append(MapDataset(name=d0.name, models=d0.models))
-                datasets.remove(d0)
-            self._datasets = datasets_list
-            self._actors = actors
-
-    def insert(self, idx, dataset):
-        from .map import MapDataset, MapDatasetActor
-
-        if isinstance(dataset, Dataset):
-            if dataset.name in self.names:
-                raise (ValueError("Dataset names must be unique"))
-            self._datasets.insert(
-                idx, MapDataset(name=dataset.name, models=dataset.models)
-            )
-
-            self._actors.insert(idx, MapDatasetActor.remote(dataset))
-        else:
-            raise TypeError(f"Invalid type: {type(dataset)!r}")
-
-    def __getattr__(self, attr):
-        """get attribute from remote each dataset"""
-
-        def wrapper(update_remote=False, **kwargs):
-            if update_remote:
-                self._update_remote_models()
-            results = ray.get([a.get_attr.remote(attr) for a in self._actors])
-            return [res(**kwargs) if inspect.ismethod(res) else res for res in results]
-
-        return wrapper
-
-    def _update_remote_models(self):
-        args = [list(d.models) for d in self._datasets]
-        ray.get([a.set_models.remote(arg) for a, arg in zip(self._actors, args)])
-
-    def stat_sum(self):
-        """Compute joint likelihood"""
-        args = [d.models.parameters.get_parameter_values() for d in self._datasets]
-        ray.get(
-            [a.set_parameter_values.remote(arg) for a, arg in zip(self._actors, args)]
-        )
-        # blocked until set_parameters_factors on actors complete
-        res = ray.get([a.stat_sum.remote() for a in self._actors])
-        return np.sum(res)
